requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
